--- template/control.js
@@ -1,29 +1,0 @@
-new (class Control {
-    #viewA;
-    #viewB;
-
-    constructor() {
-        d3.csv('../clean_vehicles.csv', (data) => {
-            // Add new div to body
-            const root = d3.select('body').append('div')
-            .style('position','fixed')
-            .style('top','0')
-            .style('left','0')
-            .style('width', '100%')
-            .style('height', '100%');
-
-            this.#viewA = new ViewA(this, root, data);
-            this.#viewB = new ViewB(this, root, data);
-        });
-    }
-
-    Test(str) {
-        console.log(str);
-    }
-
-    //Let other views know that a view has been clicked
-    Message(str){
-        this.#viewA.Hear(str)
-        this.#viewB.Hear(str)
-    }
-})()
