+++ template/public/index.html
@@ -18,6 +18,16 @@
 <body>
     <script src="control.js"></script>
     <canvas id="myChart"></canvas>
+    <script>
+        d3.csv('clean_vehicles.csv', (data) => {
+            console.log(data)
+        });
+        fetch('start')
+        .then(res=>res.json())
+        .then(data=>{
+            const con=new Control(data)
+        })
+    </script>
 </body>
 
 </html>
