+++ template/public/viewB.js
